add detail level comments
